+++ src/error.rs
@@ -1,4 +1,3 @@
-use crate::id::NamespaceId;
 use crate::xotdata::Node;
 
 /// Xot errors
@@ -30,7 +29,7 @@ pub enum Error {
     /// Can occur during serialization if a namespace is used that has no
     /// prefix is declared. Use [`Xot::create_missing_prefixes`](crate::xotdata::Xot::create_missing_prefixes)
     /// to fix this.
-    MissingPrefix(NamespaceId),
+    MissingPrefix(String),
     /// It's not allowed to include a namespace prefix in a processing instruction
     /// target name.
     NamespaceInProcessingInstruction,